JMX set attribute PoC
